refactor(app): organisms/CommandText->molecules/Command
The CommandText organism oesn't use any molecules and doesn't use any
hooks; it's weird that it's an organism. It also doesn't do as much as
it could, presenting only the way to get the actual text of a command.
If we want to add more reusable visual elements for representing
commands, they should really live alongside, and they shouldn't be a
full organism unless they're using other molecules.

So, let's move it to molecules; and let's call it Command, with an index
reexport so that other things can still grab the CommandText directly,
and we'll add further elements here.
