+++ app/src/molecules/Command/CommandText.tsx
