+++ app/src/molecules/Command/LoadCommandText.tsx
