+++ app/src/molecules/Command/PipettingCommandText.tsx
