+++ app/src/molecules/Command/TemperatureCommandText.tsx
