+++ app/src/molecules/Command/__fixtures__/mockRobotSideAnalysis.json
