+++ app/src/molecules/Command/__tests__/CommandText.test.tsx
