+++ app/src/molecules/Command/index.ts
@@ -1,0 +1,3 @@
+import {CommandText} from 'CommandText'
+
+export { CommandText }
