+++ app/src/molecules/Command/utils/getAddressableAreaDisplayName.ts
