+++ app/src/molecules/Command/utils/getCommandTextData.ts
