+++ app/src/molecules/Command/utils/getLiquidDisplayName.ts
