+++ app/src/molecules/Command/utils/getPipetteNameOnMount.ts
