+++ app/src/molecules/Command/utils/getWellRange.ts
