+++ app/src/molecules/Command/utils/index.ts
