+++ app/src/organisms/OnDeviceDisplay/RunningProtocol/CurrentRunningProtocolCommand.tsx
@@ -19,9 +19,9 @@ import {
 } from '@opentrons/components'
 import { RUN_STATUS_RUNNING, RUN_STATUS_IDLE } from '@opentrons/api-client'
 
-import { CommandText } from '../../CommandText'
+import { CommandText } from '../../../molecules/Command'
 import { RunTimer } from '../../Devices/ProtocolRun/RunTimer'
-import { getCommandTextData } from '../../CommandText/utils/getCommandTextData'
+import { getCommandTextData } from '../../../molecules/Command/utils/getCommandTextData'
 import { PlayPauseButton } from './PlayPauseButton'
 import { StopButton } from './StopButton'
 import { ANALYTICS_PROTOCOL_RUN_ACTION } from '../../../redux/analytics'
